refactor out create_tarfile
